fix: created seperate files for interfaces and improved  naming convention
