+++ src/app/core/services/appointment.service.ts
@@ -2,51 +2,8 @@ import { Injectable } from '@angular/core';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Observable } from 'rxjs';
 import { environment } from '../../../environments/environment';
+import { ApiResponse, Appointment, CreateAppointmentData, Doctor } from '../../shared/models/appointment.model';
 
-export interface Doctor {
-  id: string;
-  name: string;
-  specialization: string;
-  img?: string;
-}
-
-export interface Patient {
-  first_name: string;
-  last_name: string;
-  gender: string;
-  phone: string;
-  address: string;
-  date_of_birth: string;
-}
-
-export interface Appointment {
-  id: number;
-  patient_id: string;
-  doctor_id: string;
-  appointment_date: string;
-  time: string;
-  status: 'PENDING' | 'SCHEDULED' | 'CANCELLED' | 'COMPLETED';
-  type: string;
-  note?: string;
-  reason?: string;
-  doctor: Doctor;
-  created_at: string;
-  patient: Patient;
-}
-
-export interface CreateAppointmentData {
-  doctor_id: string;
-  appointment_date: string;
-  time: string;
-  type: string;
-  note?: string;
-}
-
-export interface ApiResponse<T> {
-  success: boolean;
-  message?: string;
-  data: T;
-}
 
 @Injectable({
   providedIn: 'root',
