+++ src/app/features/patient/appointments/appointment-detail/appointment-detail.component.ts
@@ -17,9 +17,9 @@ export class AppointmentDetailComponent {
     const today = new Date();
     const birthDate = new Date(dateOfBirth);
     let age = today.getFullYear() - birthDate.getFullYear();
-    const m = today.getMonth() - birthDate.getMonth();
+    const monthDifference = today.getMonth() - birthDate.getMonth();
     
-    if (m < 0 || (m === 0 && today.getDate() < birthDate.getDate())) {
+    if (monthDifference < 0 || (monthDifference === 0 && today.getDate() < birthDate.getDate())) {
       age--;
     }
